Add delay to write/erase operations
Other than being a TODO in the code, this fixes write errors on an STM32
board, caused by racy access to the spi status bit.

On write/erase operations, the SPI flash will set a BUSY status bit to
indicate the the operations are in progress.  We observed that at times
write or erase operations would return immediately, and that the BUSY
status bit would be set *after* the return from the call.  Adding the
delay to the loop solved that problem.