Do not exit write_enable until WEL bit is set
Any write or erase command will be ignored unless this bit is set, so
make sure the WEL bit is set before exiting write_enable().
The WEL bit is set in response to the Write Enable command, but prior to
this change, the WEL bit status was not being checked.  This would
result in occasional erase failures.